Move methods not specific to organism into the chadoFieldItemBase.
