+++ tripal_chado/src/Plugin/Field/FieldType/OBIOrganismItem.php
@@ -78,8 +78,6 @@ class OBIOrganismItem extends ChadoFieldItemBase {
    */
   public static function defaultStorageSettings() {
     $settings = parent::defaultStorageSettings();
-    $settings['max_length'] = 255;
-    $settings['tripal_custom_storage'] = 'chado';
 
     // -- Chado Table.
     // The table in Chado that the field maps to.
@@ -188,53 +186,6 @@ class OBIOrganismItem extends ChadoFieldItemBase {
   /**
    * {@inheritdoc}
    */
-  public static function schema(FieldStorageDefinitionInterface $field_definition) {
-    $schema = [
-      'columns' => [
-        'value' => [
-          'type' => 'text',
-        ],
-        'chado_schema' => [
-          'type' => 'varchar',
-          'description' => 'The name of the chado schema this record resides in.',
-          'length' => 255,
-        ],
-        'record_id' => [
-          'type' => 'int',
-          'description' => 'The primary key of this record in the chado.',
-          'size' => 'big',
-        ],
-      ],
-    ];
-
-    return $schema;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getConstraints() {
-    $constraints = parent::getConstraints();
-
-    $constraint_manager = \Drupal::typedDataManager()->getValidationConstraintManager();
-    $constraints[] = $constraint_manager->create('ComplexData', [
-      'chado_schema' => [
-        'Length' => [
-          'max' => 255,
-          'maxMessage' => t('%name: The name of the chado schema this record is associated with may not be longer than @max characters.', [
-            '%name' => $this->getFieldDefinition()->getLabel(),
-            '@max' => 255
-          ]),
-        ],
-      ],
-    ]);
-
-    return $constraints;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
   public static function generateSampleValue(FieldDefinitionInterface $field_definition) {
 
     // @todo make this specific to this field.
@@ -242,46 +193,4 @@ class OBIOrganismItem extends ChadoFieldItemBase {
     $values['value'] = $random->word(mt_rand(1, 255));
     return $values;
   }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function storageSettingsForm(array &$form, FormStateInterface $form_state, $has_data) {
-    $elements = [];
-
-    $elements['max_length'] = [
-      '#type' => 'number',
-      '#title' => t('Maximum length'),
-      '#default_value' => $this->getSetting('max_length'),
-      '#required' => TRUE,
-      '#description' => t('The maximum length of the field in characters.'),
-      '#min' => 1,
-      '#disabled' => $has_data,
-    ];
-
-    return $elements;
-  }
-
-  /**
-   * @{inheritdoc}
-   */
-  public function getValue() {
-    return $this->value;
-  }
-
-  /**
-   * @{inheritdoc}
-   */
-  public function setValue($values, $notify = TRUE) {
-    parent::setValue($values, $notify);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function isEmpty() {
-    $value = $this->get('value')->getValue();
-    return is_int($value) && $value > 0;
-  }
-
 }
